move panic file
